requirements.txt
aiofile==3.7.4
aiofiles==0.8.0
aiohttp==3.7.4.post0
alembic==1.0.0
aliyun-python-sdk-core==2.13.15
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-dysmsapi==1.0.0
allure-pytest==2.8.16
allure-python-commons==2.8.16
amqp==2.5.2
aniso8601==4.0.1
apipkg==1.5
appdirs==1.4.4
Appium-Python-Client==1.0.1
appnope==0.1.0
argon2-cffi==21.1.0
asgiref==3.4.1
asn1crypto==0.24.0
async-generator==1.10
attrs==19.3.0
Automat==20.2.0
Babel==2.10.3
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
bidict==0.21.2
billiard==3.5.0.5
bleach==4.1.0
blinker==1.4
Brotli==1.0.9
bs4==0.0.1
caio==0.9.6
celery==4.1.1
certifi==2022.9.24
cffi==1.14.4
chardet==3.0.4
charset-normalizer==2.1.1
click==7.1.2
colorlog==4.6.2
ConfigArgParse==1.2.3
configparser==3.5.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
crypto==1.4.1
cryptography==37.0.4
cssselect==1.1.0
dash==2.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-renderer==1.9.1
dash-table==5.0.0
ddt==1.6.0
decorator==4.4.2
defusedxml==0.5.0
diff-match-patch==20200713
Django==2.2
django-bootstrap4==3.0.1
django-cors-headers==2.1.0
django-crispy-forms==1.7.2
django-filter==2.0.0
django-formtools==2.1
django-import-export==1.0.0
django-plotly-dash==1.6.5
django-rest-swagger==2.1.2
django-reversion==2.0.13
django-suit==2.0a1
djangorestframework==3.7.7
dnspython==1.16.0
dominate==2.3.4
dpd-components==0.1.0
dpd-static-support==0.0.5
dukpy==0.2.2
elastic-transport==8.4.0
elasticsearch==8.4.3
email-validator==1.2.1
entrypoints==0.3
et-xmlfile==1.0.1
eventlet==0.24.1
execnet==1.7.1
Faker==13.15.1
ffmpeg==1.4
filelock==3.4.1
Flask==1.1.2
Flask-Admin==1.5.2
Flask-Avatars==0.2.2
Flask-Babel==2.0.0
Flask-BasicAuth==0.2.0
Flask-Bootstrap==3.3.7.1
Flask-Compress==1.10.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-Moment==0.11.0
flask-paginate==0.8.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SocketIO==5.1.0
Flask-SQLAlchemy==2.3.2
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.3
future==0.17.1
gevent==22.10.1
greenlet==1.1.3.post0
guess-language-spirit==0.5.3
gunicorn==20.1.0
h11==0.13.0
hvac==0.5.0
hyperlink==21.0.0
idna==2.6
idna-ssl==1.1.0
importlib-metadata==1.7.0
incremental==21.3.0
iniconfig==1.0.1
ipykernel==5.5.5
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.4
itemadapter==0.7.0
itemloaders==1.0.4
itsdangerous==1.1.0
itypes==1.2.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
jira==1.0.15
jmespath==0.9.5
jsonpath==0.82
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.2
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyter-plotly-dash==0.4.3
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.1
kombu==4.2.0
langdetect==1.0.9
locustio==0.8.1
lxml==4.9.1
Mako==1.0.12
MarkupSafe==1.1.1
mistune==0.8.4
mock==4.0.2
monotonic==1.5
more-itertools==8.3.0
msgpack==1.0.0
msgpack-python==0.5.6
multi-key-dict==2.0.3
multidict==5.1.0
Naked==0.1.31
nbclient==0.5.4
nbconvert==6.0.7
nbformat==5.1.3
nbserverproxy==0.8.8
nest-asyncio==1.5.1
notebook==6.4.3
numpy==1.21.6
oauthlib==3.1.0
openapi-codec==1.3.2
opencv-python==4.6.0.66
openpyxl==3.0.5
packaging==20.4
paho-mqtt==1.3.1
pandocfilters==1.5.0
paramiko==2.7.2
parsel==1.6.0
parso==0.7.1
pbr==5.5.1
peewee==3.9.6
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
plotly==5.3.1
pluggy==0.13.1
postgis==1.0.4
prettytable==3.5.0
prometheus-client==0.11.0
prompt-toolkit==3.0.7
Protego==0.2.1
protobuf==3.7.1
ptyprocess==0.6.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.9.7
PyDispatcher==2.0.5
pyecharts==1.9.1
pyecharts-jupyter-installer==0.0.3
pyee==8.2.2
Pygments==2.7.0
PyJWT==2.4.0
pymongo==3.6.0
PyMySQL==0.8.1
PyNaCl==1.4.0
pyOpenSSL==22.0.0
pyparsing==2.4.7
pypinyin==0.36.0
pyppeteer==1.0.2
pyrsistent==0.18.0
PySnooper==0.4.1
pytest==5.4.3
pytest-emoji==0.2.0
pytest-forked==1.1.3
pytest-html==2.1.1
pytest-metadata==1.9.0
pytest-ordering==0.6
pytest-rerunfailures==9.0
pytest-sugar==0.9.3
pytest-xdist==1.32.0
python-crontab==2.2.8
python-dateutil==2.8.0
python-docx==0.8.6
python-dotenv==0.20.0
python-editor==1.0.4
python-engineio==4.2.0
python-jenkins==1.7.0
python-memcached==1.59
python-socketio==5.3.0
pytz==2017.3
PyYAML==5.3.1
pyzmq==19.0.1
qiniu==7.2.6
qtconsole==5.1.1
QtPy==1.11.0
queuelib==1.6.2
redis==3.5.3
redis-py-cluster==1.3.4
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.0
rocketmq==0.3.11
rq==1.2.0
schedule==1.1.0
Scrapy==2.6.2
selenium==3.141.0
Send2Trash==1.8.0
service-identity==21.1.0
shellescape==3.4.1
shortuuid==0.5.0
simplejson==3.16.0
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.3.5
sqlparse==0.4.2
stripe==2.10.1
supervisor==4.2.4
tablib==3.1.0
tenacity==8.0.1
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
thrift==0.11.0
timeloop==1.0.2
tldextract==3.1.2
toml==0.10.1
tornado==6.1
tqdm==4.64.1
traitlets==4.3.3
Twisted==21.2.0
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.3
uvicorn==0.16.0
vine==1.3.0
visitor==0.1.3
w3lib==1.22.0
wcwidth==0.2.4
webencodings==0.5.1
websockets==10.4
Werkzeug==1.0.1
Whoosh==2.7.4
widgetsnbextension==3.5.1
WTForms==2.3.3
xlrd==1.2.0
XlsxWriter==1.0.5
xlwt==1.3.0
yarl==1.6.3
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
